make_dreamcatcher/requirements.txt
numpy
pillow
opencv-python
shapely
trimesh

# to fix error: `ValueError: No available triangulation engine!`
manifold3d